Add network utils and test in service id network is reachable

+++ src/com/akelio/android/acollab/service/ActivityStreamService.java
@@ -16,6 +16,7 @@ import android.util.Log;
 import com.akelio.android.acollab.contract.ActivityContract;
 import com.akelio.android.acollab.db.DbHelper;
 import com.akelio.android.acollab.entity.ActivityItem;
+import com.akelio.android.acollab.utils.NetworkUtils;
 
 public class ActivityStreamService extends IntentService {
 
@@ -31,48 +32,52 @@ public class ActivityStreamService extends IntentService {
 	protected void onHandleIntent(Intent intent) {
 		Log.d(TAG, "activityStreamService launched");
 		System.out.println("activityStreamService launched");
-		try {
-			HttpAuthentication authHeader = new HttpBasicAuthentication("admin", "admin");
-			HttpHeaders requestHeaders = new HttpHeaders();
-			requestHeaders.setAuthorization(authHeader);
-			HttpEntity<?> requestEntity = new HttpEntity<Object>(requestHeaders);
 
-			RestTemplate restTemplate = new RestTemplate();
-			restTemplate = new RestTemplate();
-			restTemplate.getMessageConverters().add(new GsonHttpMessageConverter());
-			DbHelper dbHelper = new DbHelper(this);
-			SQLiteDatabase db = dbHelper.getWritableDatabase();
-			ContentValues values = new ContentValues();
+		if (NetworkUtils.isNetworkReachable(getApplicationContext())) {
+			try {
+				HttpAuthentication authHeader = new HttpBasicAuthentication("admin", "admin");
+				HttpHeaders requestHeaders = new HttpHeaders();
+				requestHeaders.setAuthorization(authHeader);
+				HttpEntity<?> requestEntity = new HttpEntity<Object>(requestHeaders);
 
-			ResponseEntity<ActivityItem[]> res = restTemplate.exchange(URL, HttpMethod.GET, requestEntity, ActivityItem[].class);
-			ActivityItem[] ulis = res.getBody();
+				RestTemplate restTemplate = new RestTemplate();
+				restTemplate = new RestTemplate();
+				restTemplate.getMessageConverters().add(new GsonHttpMessageConverter());
+				DbHelper dbHelper = new DbHelper(this);
+				SQLiteDatabase db = dbHelper.getWritableDatabase();
+				ContentValues values = new ContentValues();
 
-			db.execSQL("delete from " + ActivityContract.TABLE);
+				ResponseEntity<ActivityItem[]> res = restTemplate.exchange(URL, HttpMethod.GET, requestEntity, ActivityItem[].class);
+				ActivityItem[] ulis = res.getBody();
 
-			int i = 0;
-			for (i = 0; i < ulis.length; i++) {
-				ActivityItem a = ulis[i];
-				values.clear(); //
-				values.put(ActivityContract.Column.ID, a.getActivityId());
-				values.put(ActivityContract.Column.TENANT_ID, a.getTenantId());
-				values.put(ActivityContract.Column.USER_ID, a.getUserId());
-				values.put(ActivityContract.Column.SPACE_ID, a.getUserId());
-				values.put(ActivityContract.Column.SPACE_NAME, a.getSpaceName());
-				values.put(ActivityContract.Column.TITLE, a.getTitle());
-				values.put(ActivityContract.Column.USER_NAME, a.getUserName());
-				values.put(ActivityContract.Column.DATE_CREATED, a.getDateCreated());
-				values.put(ActivityContract.Column.MODULE_ID, a.getModuleId());
-				values.put(ActivityContract.Column.MODULE_TYPE, a.getModuleType());
-				values.put(ActivityContract.Column.DATE_CREATED, a.getDateCreated());
+				db.execSQL("delete from " + ActivityContract.TABLE);
 
-				db.insertWithOnConflict(ActivityContract.TABLE, null, values, SQLiteDatabase.CONFLICT_REPLACE);
+				int i = 0;
+				for (i = 0; i < ulis.length; i++) {
+					ActivityItem a = ulis[i];
+					values.clear(); //
+					values.put(ActivityContract.Column.ID, a.getActivityId());
+					values.put(ActivityContract.Column.TENANT_ID, a.getTenantId());
+					values.put(ActivityContract.Column.USER_ID, a.getUserId());
+					values.put(ActivityContract.Column.SPACE_ID, a.getUserId());
+					values.put(ActivityContract.Column.SPACE_NAME, a.getSpaceName());
+					values.put(ActivityContract.Column.TITLE, a.getTitle());
+					values.put(ActivityContract.Column.USER_NAME, a.getUserName());
+					values.put(ActivityContract.Column.DATE_CREATED, a.getDateCreated());
+					values.put(ActivityContract.Column.MODULE_ID, a.getModuleId());
+					values.put(ActivityContract.Column.MODULE_TYPE, a.getModuleType());
+					values.put(ActivityContract.Column.DATE_CREATED, a.getDateCreated());
+
+					db.insertWithOnConflict(ActivityContract.TABLE, null, values, SQLiteDatabase.CONFLICT_REPLACE);
+				}
+				System.out.println("Nb activity : " + i);
+				db.close();
+			} catch (Exception e) {
+				e.printStackTrace();
 			}
-			System.out.println("Nb activity : " + i);
-			db.close();
-		} catch (Exception e) {
-			e.printStackTrace();
+		} else {
+			Log.d(TAG, "activityStreamService no network reachable");
 		}
-
 	}
 
 	@Override
